Removed unnecessary casts

+++ jenjin-core-client/src/main/java/com/jenjinstudios/client/net/LoginTracker.java
@@ -75,7 +75,7 @@ public class LoginTracker
     public boolean sendLoginRequestAndWaitForResponse() {
         sendLoginRequest();
         long startTime = System.currentTimeMillis();
-        while (waitingForResponse && ((System.currentTimeMillis() - startTime) < (long) MILLIS_IN_30_SECONDS))
+        while (waitingForResponse && ((System.currentTimeMillis() - startTime) < MILLIS_IN_30_SECONDS))
         {
             waitTenMillis();
         }
@@ -88,7 +88,7 @@ public class LoginTracker
     public void sendLogoutRequestAndWaitForResponse() {
         sendLogoutRequest();
         long startTime = System.currentTimeMillis();
-        while (waitingForResponse && ((System.currentTimeMillis() - startTime) < (long) MILLIS_IN_30_SECONDS))
+        while (waitingForResponse && ((System.currentTimeMillis() - startTime) < MILLIS_IN_30_SECONDS))
         {
             waitTenMillis();
         }
